requirements.txt
Python 3.12.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2

Virtual Environment 
python -m venv venv
source ./venv/Scripts/activate  # Activate it